Move renderer ANR into an integration
