+++ src/main/sdk.ts
@@ -26,7 +26,6 @@ import {
 import { Session, session, WebContents } from 'electron';
 
 import { IPCMode } from '../common/ipc';
-import { configureRendererAnr } from './anr';
 import { getDefaultEnvironment, getDefaultReleaseName, getSdkInfo } from './context';
 import { additionalContextIntegration } from './integrations/additional-context';
 import { childProcessIntegration } from './integrations/child-process';
@@ -38,6 +37,7 @@ import { electronNetIntegration } from './integrations/net-breadcrumbs';
 import { normalizePathsIntegration } from './integrations/normalize-paths';
 import { onUncaughtExceptionIntegration } from './integrations/onuncaughtexception';
 import { preloadInjectionIntegration } from './integrations/preload-injection';
+import { rendererAnrIntegration } from './integrations/renderer-anr';
 import { rendererProfilingIntegration } from './integrations/renderer-profiling';
 import { screenshotsIntegration } from './integrations/screenshots';
 import { sentryMinidumpIntegration } from './integrations/sentry-minidump';
@@ -60,6 +60,7 @@ export function getDefaultIntegrations(options: ElectronMainOptions): Integratio
     additionalContextIntegration(),
     screenshotsIntegration(),
     gpuContextIntegration(),
+    rendererAnrIntegration(),
 
     // Main process sessions
     mainProcessSessionIntegration(),
@@ -135,12 +136,6 @@ export interface ElectronMainOptionsInternal
    * Enables injection of 'js-profiling' document policy headers and ensure profiles are forwarded with transactions
    */
   enableRendererProfiling?: boolean;
-
-  /**
-   * Enables injection of 'include-js-call-stacks-in-crash-reports' document policy headers so that renderer call stacks
-   * can be captured from the main process
-   */
-  enableRendererStackCapture?: boolean;
 }
 
 // getSessions and ipcMode properties are optional because they have defaults
@@ -181,9 +176,7 @@ export function init(userOptions: ElectronMainOptions): void {
   }
 
   removeRedundantIntegrations(options);
-  configureIPC(options);
   configureUtilityProcessIPC();
-  configureRendererAnr(options);
 
   setNodeAsyncContextStrategy();
 
@@ -199,6 +192,8 @@ export function init(userOptions: ElectronMainOptions): void {
 
   scope.setClient(client);
   client.init();
+
+  configureIPC(client, options);
 
   // If users opt-out of this, they _have_ to set up OpenTelemetry themselves
   // There is no way to use this SDK without OpenTelemetry!
