+++ test/e2e/test-apps/anr/anr-renderer/src/main.js
@@ -1,12 +1,12 @@
 const path = require('path');
 
 const { app, BrowserWindow } = require('electron');
-const { init } = require('@sentry/electron/main');
+const { init, rendererAnrIntegration } = require('@sentry/electron/main');
 
 init({
   dsn: '__DSN__',
   debug: true,
-  enableRendererStackCapture: true,
+  integrations: [rendererAnrIntegration({ injectDocumentPolicy: true })],
   onFatalError: () => {},
 });
